server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "fastapi[standard]>=0.120.4",
    "passlib>=1.7.4",
    "pgvector>=0.4.1",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "gunicorn>=23.0.0",
    "uvicorn[standard]>=0.38.0",
    "psycopg2-binary>=2.9.11",
    "aiofiles>=25.1.0",
    "supabase>=2.24.0",
    "asyncpg>=0.30.0",
]
